Add integration and validation tests for TACACS+ configuration parsing
- Implemented integration tests for mapping TACACS+ server configurations to connection settings, including inline TLS material, PSK, and reference-based TLS servers.
- Added validation tests to ensure correct parsing and error handling for various TACACS+ configurations, including duplicate addresses, missing required fields, and invalid types.
- Enhanced test coverage for credential references and security configurations, ensuring compliance with expected behaviors and error messages.
